LPS-205346 SF, rename: benchmark -> benchmarks
