--- modules/util/portal-tools-benchmark-test/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal Tools Benchmark Test
-Bundle-SymbolicName: com.liferay.portal.tools.benchmark.test
-Bundle-Version: 1.0.0
