--- modules/util/portal-tools-benchmark-test/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
-	testIntegrationImplementation project(":util:portal-tools-benchmark")
-}
