--- modules/util/portal-tools-benchmark/bnd.bnd
@@ -1,8 +1,0 @@
-Bundle-Name: Liferay Portal Tools Benchmark
-Bundle-SymbolicName: com.liferay.portal.tools.benchmark
-Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.portal.tools.benchmark,\
-	com.liferay.portal.tools.benchmark.task
-Import-Package: *;resolution:=optional
-Main-Class: com.liferay.portal.tools.benchmark.BenchmarkTestLauncher
