--- modules/util/portal-tools-benchmark/build.gradle
@@ -1,23 +1,0 @@
-dependencies {
-	compileInclude group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileInclude group: "com.mysql", name: "mysql-connector-j", version: "8.0.33"
-	compileInclude project(":core:petra:petra-function")
-	compileInclude project(":core:petra:petra-lang")
-	compileInclude project(":core:petra:petra-string")
-}
-
-test {
-	systemProperties(
-		System.properties.subMap([
-			"benchmark.test.excluded.companies",
-			"benchmark.test.jdbc.driverClassName",
-			"benchmark.test.jdbc.password",
-			"benchmark.test.jdbc.url",
-			"benchmark.test.jdbc.username",
-			"benchmark.test.run.count",
-			"benchmark.test.skip.warm.up",
-			"benchmark.test.thread.count",
-			"benchmark.test.user.password"
-		])
-	)
-}
