--- modules/util/portal-tools-benchmark/src/main/java/com/liferay/portal/tools/benchmark/task/BenchmarkTask.java
@@ -1,18 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.tools.benchmark.task;
-
-import java.util.List;
-import java.util.Map;
-
-/**
- * @author Tina Tian
- */
-public interface BenchmarkTask {
-
-	public List<Map.Entry<String, Long>> execute() throws Exception;
-
-}
