--- modules/util/portal-tools-benchmark/src/test/java/com/liferay/portal/tools/benchmark/BenchmarkTestTest.java
@@ -1,31 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.tools.benchmark;
-
-import com.liferay.portal.test.rule.LiferayUnitTestRule;
-
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-
-/**
- * @author Tina Tian
- */
-public class BenchmarkTestTest {
-
-	@ClassRule
-	@Rule
-	public static final LiferayUnitTestRule liferayUnitTestRule =
-		LiferayUnitTestRule.INSTANCE;
-
-	@Test
-	public void testLogin() throws Exception {
-		BenchmarkTest benchmarkTest = new BenchmarkTest();
-
-		benchmarkTest.login();
-	}
-
-}
