+++ modules/util/portal-tools-benchmarks-test/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Tools Benchmarks Test
+Bundle-SymbolicName: com.liferay.portal.tools.benchmarks.test
+Bundle-Version: 1.0.0
