+++ modules/util/portal-tools-benchmarks/src/main/java/com/liferay/portal/tools/benchmarks/task/BenchmarksTask.java
@@ -1,0 +1,18 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.tools.benchmarks.task;
+
+import java.util.List;
+import java.util.Map;
+
+/**
+ * @author Tina Tian
+ */
+public interface BenchmarksTask {
+
+	public List<Map.Entry<String, Long>> execute() throws Exception;
+
+}
