+++ modules/util/portal-tools-benchmarks/src/main/resources/com/liferay/portal/tools/benchmarks/packageinfo
