+++ modules/util/portal-tools-benchmarks/src/test/java/com/liferay/portal/tools/benchmarks/BenchmarksTestTest.java
@@ -1,0 +1,31 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.tools.benchmarks;
+
+import com.liferay.portal.test.rule.LiferayUnitTestRule;
+
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+
+/**
+ * @author Tina Tian
+ */
+public class BenchmarksTestTest {
+
+	@ClassRule
+	@Rule
+	public static final LiferayUnitTestRule liferayUnitTestRule =
+		LiferayUnitTestRule.INSTANCE;
+
+	@Test
+	public void testLogin() throws Exception {
+		BenchmarksTest benchmarksTest = new BenchmarksTest();
+
+		benchmarksTest.login();
+	}
+
+}
