+++ test.properties
@@ -1119,7 +1119,7 @@
         modules/apps/static/portal-lpkg-deployer/**,\
         modules/test/external-data-source-test/**,\
         modules/test/semantic-versioning-test/**,\
-        modules/util/portal-tools-benchmark/**,\
+        modules/util/portal-tools-benchmarks/**,\
         portal-kernel/test/unit/com/liferay/portal/log/assertor/PortalLogAssertorTest.java,\
         portal-kernel/test/unit/com/liferay/portal/modules/ModulesStructureTest.java
